pyproject.toml
[project]
name = "wdtagger"
version = "0.13.0"
description = "A simple and easy-to-use wrapper for the tagger model created by [SmilingWolf](https://github.com/SmilingWolf) which is specifically designed for tagging anime illustrations."
authors = [{ name = "Jianqi Pan", email = "jannchie@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface-hub>=0.26.2",
    "numpy<2",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
]

[dependency-groups]
dev = ["pytest>=8.3.3", "pytest-benchmark>=5.1.0", "ruff>=0.8.0"]


[project.optional-dependencies]
cpu = ["torch>=2.0.0", "torchvision>=0.20.1", "timm>=1.0.11"]
gpu = ["torch>=2.0.0", "torchvision>=0.20.1", "timm>=1.0.11"]

[tool.ruff]
line-length = 140
select = ["ALL"]

ignore = [
    "PGH",
    "RUF003",
    "BLE001",
    "ERA001",
    "FIX002",
    "TD002",
    "TD003",
    "D",
    "N812",
]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = [
    # at least this three should be fine in tests:
    "S101",    # asserts allowed in tests
    "PLR2004", # Magic value used in comparison
]
[tool.pyright]

[tool.uv]
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" },
        { extra = "cuda11" },
        { extra = "cuda12" },
    ],
]
package = true

[tool.uv.sources]
torch = [
    { index = "torch-cpu", extra = "cpu" },
    { index = "torch-gpu", extra = "gpu" },
    { index = "torch-cuda11", extra = "cuda11" },
    { index = "torch-cuda12", extra = "cuda12" },
]

torchvision = [
    { index = "torch-cpu", extra = "cpu" },
    { index = "torch-gpu", extra = "gpu" },
    { index = "torch-cuda11", extra = "cuda11" },
    { index = "torch-cuda12", extra = "cuda12" },
]

[[tool.uv.index]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "torch-cuda11"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[[tool.uv.index]]
name = "torch-cuda12"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
